+++ src/InternalBenchmarks/LexTimeBenchmark.cs
@@ -1,5 +1,4 @@
 ﻿using BenchmarkDotNet.Attributes;
-using BenchmarkDotNet.Engines;
 using BenchmarkDotNet.Jobs;
 using Loretta.CodeAnalysis;
 using Loretta.CodeAnalysis.Lua;
@@ -14,7 +13,7 @@ namespace Loretta.InternalBenchmarks
     [MemoryDiagnoser]
     public class LexTimeBenchmark
     {
-        private static readonly LuaParseOptions _parseOptions = new(LuaSyntaxOptions.All);
+        private static readonly LuaParseOptions s_parseOptions = new(LuaSyntaxOptions.All);
 
         [ParamsSource(nameof(Files))]
         public TestFile File { get; set; }
@@ -25,11 +24,12 @@ namespace Loretta.InternalBenchmarks
             {
                 yield return TestFile.Load("samples/benchies/anim.lua");
                 yield return TestFile.Load("samples/benchies/rustic.lua");
+                yield return TestFile.Load("samples/benchies/rustic-24mb.lua");
             }
         }
 
         [Benchmark]
         public ImmutableArray<SyntaxToken> Lex() =>
-            SyntaxFactory.ParseTokens(File.Text, options: _parseOptions).ToImmutableArray();
+            SyntaxFactory.ParseTokens(File.Text, options: s_parseOptions).ToImmutableArray();
     }
 }